+++ src/helpers/StanzaManager.ts
@@ -1,6 +1,7 @@
 import * as XMPP from 'stanza';
 import { PlayerContextInfo, RoomDetails } from '../App';
 import { SubsManager } from './SubscriptionManager';
+import { JSONItem } from 'stanza/protocol';
 
 /** class that handles subscriptions to pub-sub nodes, supporting callbacks
 * for when those documents change
@@ -107,5 +108,48 @@ export class StanzaManager {
     } else {
       throw new Error('No client')
     }
+  }
+  async createNodeIfNecessaryThenPublish ( node: string,
+    description: string, jsonItem: JSONItem, allVersions = false
+  ): Promise<void> {
+    // check if node exists
+    this.client.getNodeConfig(this.pubJid, node).then(() => {
+      // success
+      // console.log('got config for node', node)
+    }).catch((err: { error: { condition: string } }) => {
+      // ignore compiler warning for unknown type on line below
+      if (err.error.condition === 'item-not-found') {
+        console.log('creating node', node)
+        // create node
+        this.client.createNode(this.pubJid, node).then(() => {
+          console.log('created node', node)
+          // now configure it
+          const nodeConfig: XMPP.Stanzas.DataForm = {
+            fields:[
+              {name: 'pubsub#description', value: description},
+              {name: 'pubsub#type', value: 'json'},
+              {name: 'pubsub#persist_items', value: true}
+            ],
+            type: 'submit',
+            title: 'Node configuration'
+          }
+          if (allVersions && nodeConfig.fields) {
+            // indicate we want to store all previous versions
+            nodeConfig.fields.push({name: 'pubsub#max_items', value: '1000'})
+          }
+          console.log('about to config node', node, nodeConfig)
+          this.client.configureNode(this.pubJid, node, nodeConfig).then(() => {
+            // success
+            console.log('new node configured')
+          }).catch((err: unknown) => {
+            console.error('Error creating/configuring node', err)
+          })
+      })}}).then(() => {
+        this.client.publish(this.pubJid, node, jsonItem).then((res) => {
+          console.log('Published game state', res)
+        })
+      }).catch((err: unknown) => {
+          console.error('Error publishing game state', err)
+      })
   }
 }
